Update README.md
This project uses Python libraries such as pandas for data extraction and creation of a data frame, sqlalchemy for DB purposes, regex for obtaining specific desired information and is deployed via Uvicorn ASGI web server.
In this stage of the project, the request is conducted by a simple HTTP request where the user types the "gemi" number of a company and displays the desired data from that company. The goal of this project was for me to obtain knowledge and experience in data extraction, refinement, storage and overall familiarize myself with the overall concept of Data Analysis.